Moved the synced attrs to its own function
